feat!: move command queuing to a command queuer driving port
