+++ src/Application/Bus/CommandDispatcher.php
@@ -12,24 +12,16 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Application\Bus;
 
-use Closure;
 use CloudCreativity\Modules\Contracts\Application\Bus\CommandHandlerContainer;
-use CloudCreativity\Modules\Contracts\Application\Ports\Driven\Queue;
 use CloudCreativity\Modules\Contracts\Application\Ports\Driving\CommandDispatcher as ICommandDispatcher;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Command;
 use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer;
 use CloudCreativity\Modules\Contracts\Toolkit\Result\Result;
 use CloudCreativity\Modules\Toolkit\Pipeline\MiddlewareProcessor;
 use CloudCreativity\Modules\Toolkit\Pipeline\PipelineBuilder;
-use RuntimeException;
 
 class CommandDispatcher implements ICommandDispatcher
 {
-    /**
-     * @var null|Queue|Closure(): Queue
-     */
-    private Queue|Closure|null $queue;
-
     /**
      * @var array<string|callable>
      */
@@ -40,14 +32,11 @@ class CommandDispatcher implements ICommandDispatcher
      *
      * @param CommandHandlerContainer $handlers
      * @param PipeContainer|null $middleware
-     * @param null|Closure(): Queue $queue
      */
     public function __construct(
         private readonly CommandHandlerContainer $handlers,
         private readonly ?PipeContainer $middleware = null,
-        ?Closure $queue = null,
     ) {
-        $this->queue = $queue;
     }
 
     /**
@@ -79,24 +68,6 @@ class CommandDispatcher implements ICommandDispatcher
         assert($result instanceof Result, 'Expecting pipeline to return a result object.');
 
         return $result;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function queue(Command $command): void
-    {
-        if ($this->queue === null) {
-            throw new RuntimeException(
-                'Commands cannot be queued because the command dispatcher has not been given a queue factory.',
-            );
-        }
-
-        if ($this->queue instanceof Closure) {
-            $this->queue = ($this->queue)();
-        }
-
-        $this->queue->push($command);
     }
 
     /**
